config: move afterConfSetup out of loadConfig
Because it touches globals, so it can't live in the independent config
package.

Also move generateRandomNodeID to main.go, since that's where the only
use is.
